% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{distOverlay2}
\alias{distOverlay2}
\title{Calculate the distance from each breakpoint to closest motif
Overlay the same number of random simulated breakpoints}
\usage{
distOverlay2(...,
  featureDir = "~/Desktop/script_test/svBreaks/rawdata/features/",
  from = "bps", lim = 2.5, n = 2, plot = TRUE, keep = NULL,
  position = "centre")
}
\description{
Calculate the distance from each breakpoint to closest motif
Overlay the same number of random simulated breakpoints
}
\keyword{motif}
