% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{EnrichmentVolcano}
\alias{EnrichmentVolcano}
\title{Plot the enrichment of SNVs in genes features}
\usage{
EnrichmentVolcano(d)
}
\description{
Plot the enrichment of SNVs in genes features
}
\keyword{enrichment}
