% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEnrichment.R
\name{featureEnrichment}
\alias{featureEnrichment}
\title{featureEnrichment}
\usage{
featureEnrichment(features = "data/genomic_features.txt")
}
\arguments{
\item{features}{File containing total genomic lengths of features [Default 'data/genomic_features.txt']}
}
\description{
Function to calculate enrichment of snv hits in genomic features
}
\keyword{enrichment}
