% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPackage.R
\name{generateData}
\alias{generateData}
\title{generateData
Prepare data for dist2motif}
\usage{
generateData(..., bed_in = NULL, annotated_file = TRUE, sim = NULL,
  keep = NULL, chromosomes = c("2L", "2R", "3L", "3R", "X"))
}
\description{
generateData
Prepare data for dist2motif
}
\keyword{simulate}
