% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTris.R
\name{genTris}
\alias{genTris}
\title{genTris}
\usage{
genTris()
}
\value{
Character string containing all 96 trinucleotides
genTris()
}
\description{
This function returns all possible trinucleotide combinations
}
\keyword{trinucleotides}
