% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(infile = "data/GW.snv.dist.txt")
}
\arguments{
\item{infile}{File to process [Required]}
}
\value{
Dataframe
}
\description{
Function to clean cnv files
}
\keyword{get}
