% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triCount.R
\name{getTriFromFasta}
\alias{getTriFromFasta}
\title{A function to calculate trinucleotide frequences from a supplied fasta file}
\usage{

  getTriFromFasta(fastaFile = "~/Documents/Curie/Data/Genomes/Dmel_v6.12/Dmel_6.12.fasta")
}
\arguments{
\item{fastaFile}{A fasta file (genome, exome)}
}
\value{
Dataframe of trinucs and freqs (or counts if count=1)
}
\description{
A function to calculate trinucleotide frequences from a supplied fasta file
}
