% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indels.R
\name{indel_lengths}
\alias{indel_lengths}
\title{indel_length}
\usage{
indel_lengths(..., exclude = TRUE,
  indels = "data/annotated_indels.txt", plot = TRUE)
}
\description{
Plot indel size per sample
}
\keyword{count}
