% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indels.R
\name{plot_indel_lengths}
\alias{plot_indel_lengths}
\title{indel_length}
\usage{
plot_indel_lengths(..., exclude = TRUE,
  indels = "data/annotated_indels.txt")
}
\description{
Plot indel size per sample
}
\keyword{count}
