% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTris.R
\name{printTris}
\alias{printTris}
\title{printTris}
\usage{
printTris()
}
\value{
Character string containing all 96 trinucleotides
printTris()
}
\description{
This function returns all possible trinucleotide combinations
}
\keyword{trinucleotides}
