% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{rainfall}
\alias{rainfall}
\title{rainfall}
\usage{
rainfall(..., snv_data = NULL, write = FALSE, title = NULL,
  exclude_chroms = c("4", "Y"))
}
\description{
Plot log10 distances between snvs as rainfall plot
}
\keyword{rainfall}
