% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPackage.R
\name{readBed}
\alias{readBed}
\title{Simple function to read in bedfiles and return as a 3 col df}
\usage{
readBed(bed_in = NULL, annotated_file = FALSE)
}
\description{
Simple function to read in bedfiles and return as a 3 col df
}
\keyword{bed}
