% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesPlot.R
\name{samplesPlot}
\alias{samplesPlot}
\title{samplesPlot}
\usage{
samplesPlot(count = NA)
}
\arguments{
\item{count}{Output total counts instead of frequency if set [Default no]}
}
\description{
Plot the snv distribution for each sample
}
\keyword{spectrum}
