% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCols.R
\name{setCols}
\alias{setCols}
\title{setCols}
\usage{
setCols(df, col)
}
\arguments{
\item{df}{Dataframe [Required]}

\item{col}{Column of dataframe. Colours will be set to levels(df$cols) [Required]}
}
\description{
Show top hit genes
}
\keyword{cols}
