% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigTypes.R
\name{sigTypes}
\alias{sigTypes}
\title{sigTypes}
\usage{
sigTypes()
}
\description{
Calculate and plot the contribution of mutational signatures accross samples using the package `deconstructSigs`
}
\keyword{signatures}
