% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triCount.R
\name{triFreq}
\alias{triFreq}
\title{triFreq}
\usage{
triFreq(genome = NA, count = NA)
}
\arguments{
\item{genome}{BS.genome file defaults to BSgenome.Dmelanogaster.UCSC.dm6}

\item{count}{Output total counts instead of frequency if set [Default no]}
}
\value{
Dataframe of trinucs and freqs (or counts if count=1)
}
\description{
This function counts the number of times each triunucleotide is found in a supplied genome
}
\keyword{trinucleotides}
