% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2motif.R
\name{dist2motif2}
\alias{dist2motif2}
\title{dist2Motif2
Calculate the distance from each breakpoint to closest motif in a directory of files}
\usage{
dist2motif2(..., breakpoints = NA, feature_file = NA,
  featureDir = "~/Desktop/script_test/svBreaks/rawdata/features/",
  sim = FALSE, keep = NULL, position = "centre")
}
\description{
dist2Motif2
Calculate the distance from each breakpoint to closest motif in a directory of files
}
\keyword{motif}
