% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEnrichment.R
\name{featureEnrichment}
\alias{featureEnrichment}
\title{featureEnrichment}
\usage{
featureEnrichment(features = "data/genomic_features.txt",
  genome_length = 118274340, write = FALSE)
}
\arguments{
\item{features}{File containing total genomic lengths of features [Default 'data/genomic_features.txt']}

\item{genome_length}{The total legnth of the genome [Default 118274340 (mappable regions on chroms 2, 3, 4, X & Y for Drosophila melanogastor Dmel6.12)]}
}
\value{
A snv_data frame with FC scores for all genes seen at least n times in snv snv_data
}
\description{
Calculate the enrichment of snv hits in genomic features
A 'features' file must be provided with the follwing format:
feature	length	percentage
This can be generated using the script 'script/genomic_features.pl' and a genome .gtf file
The defualt genome length is set to the mappable regions of the Drosophila melanogastor Dmel6.12 genome (GEM mappability score > .5)
(118274340). The full, assembled genome legnth for chroms 2/3/4/X/Y is 137547960
}
\details{
Function to calculate enrichment of snv hits in genomic features
}
\keyword{enrichment}
