% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{geneEnrichment}
\alias{geneEnrichment}
\title{geneEnrichment}
\usage{
geneEnrichment(gene_lengths = "data/gene_lengths.txt", n = 2)
}
\arguments{
\item{gene_lengths}{File containing all genes and their lengths (as generated by 'script/genomefeatures.pl') [Default 'data/gene_lengths.txt']}

\item{n}{The number of times we need to have seen a gene in our data to view its enrichment score [Default 2]}
}
\value{
A data frame with FC scores for all genes seen at least n times in snv data
}
\description{
Function to calculate fold change enrichment in a set of snv calls correcting for gene length
}
\keyword{enrichment}
