% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{geneEnrichment}
\alias{geneEnrichment}
\title{geneEnrichment}
\usage{
geneEnrichment(gene_lengths = "data/gene_lengths.txt")
}
\arguments{
\item{gene_lengths}{File containing all genes and their lengths (as generated by 'script/genomefeatures.pl')[Default 'data/gene_lengths.txt']}
}
\description{
Function to calculate fold change enrichment in a set of svc calls correcting for gene length
}
\keyword{enrichment}
