% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(..., infile = "data/annotated_snvs.txt", exclude = TRUE,
  expression_data = "data/isc_genes_rnaSeq.csv", type = "snv")
}
\arguments{
\item{infile}{File to process [Required]}
}
\value{
Dataframe
}
\description{
Function to clean cnv files
}
\keyword{get}
