% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2tss.R
\name{tssDist}
\alias{tssDist}
\title{tssDist}
\usage{
tssDist(tss_pos = "data/tss_positions.txt")
}
\arguments{
\item{tss_pos}{File containing "gene chrom tss" information [Default 'data/tss_positions.txt']}
}
\description{
Plot distance to TSS distribution
}
\keyword{tss}
